Moving the one function from GUI into main. This is vizmo application specific items for now, so they belong to the main. Also MainWin in this function is changing from a pointer to not a pointer, so fixing the destructor appropriately.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@343 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/GUI/CaptureOptions.cpp
@@ -126,7 +126,7 @@ CaptureOptions::CaptureMovie(){
 
     //for each frame, update the image, compute a filename, and save the image
     size_t frame = 0;
-    for(int i = msd.m_startFrame; i <= msd.m_endFrame; i += msd.m_stepSize, ++frame){
+    for(size_t i = msd.m_startFrame; i <= msd.m_endFrame; i += msd.m_stepSize, ++frame){
       //update progress bar
       progress.setValue(i - msd.m_startFrame);
       qApp->processEvents();
